Replace watcher service implementation
