+++ internal/service/std.go
@@ -1,11 +1,9 @@
 package service
 
 import (
-	"context"
 	"errors"
 
 	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/nokamoto/egosla/api"
 	"github.com/nokamoto/egosla/internal/mysql"
 	"go.uber.org/zap"
 	"google.golang.org/genproto/protobuf/field_mask"
@@ -56,6 +54,11 @@ func (s *std) get(validate func() error, req getRequest) (proto.Message, error) 
 	}
 
 	return res, nil
+}
+
+type listRequest interface {
+	GetPageToken() string
+	GetPageSize() int32
 }
 
 func (s *std) list(req listRequest) ([]proto.Message, string, error) {
@@ -119,6 +122,10 @@ func (s *std) update(validate func() error, req updateRequest, value proto.Messa
 	return res, nil
 }
 
+type deleteRequest interface {
+	GetName() string
+}
+
 func (s *std) delete(validate func() error, req deleteRequest) (*empty.Empty, error) {
 	l := s.logger.With(zap.String("method", "delete"), zap.String("name", req.GetName()))
 	l.Debug("delete")
@@ -134,85 +141,4 @@ func (s *std) delete(validate func() error, req deleteRequest) (*empty.Empty, er
 	}
 
 	return &empty.Empty{}, nil
-}
-
-type WatcherV2 struct {
-	api.UnimplementedWatcherServiceServer
-	std    *std
-	naming nameGenerator
-}
-
-func (w *WatcherV2) revert(v proto.Message) (*api.Watcher, error) {
-	res, ok := v.(*api.Watcher)
-	if !ok {
-		w.std.logger.Error("[bug] unknown value", zap.Any("value", v))
-		return nil, status.Error(codes.Internal, "internal error occurred")
-	}
-	return res, nil
-}
-
-func (w *WatcherV2) CreateWatcher(ctx context.Context, req *api.CreateWatcherRequest) (*api.Watcher, error) {
-	created := &api.Watcher{
-		Name:     w.naming.newName(),
-		Keywords: req.GetWatcher().GetKeywords(),
-	}
-	if err := w.std.create(func() error {
-		return nil
-	}, created); err != nil {
-		return nil, err
-	}
-	return created, nil
-}
-
-func (w *WatcherV2) GetWatcher(ctx context.Context, req *api.GetWatcherRequest) (*api.Watcher, error) {
-	res, err := w.std.get(func() error {
-		return nil
-	}, req)
-	if err != nil {
-		return nil, err
-	}
-
-	return w.revert(res)
-}
-
-func (w *WatcherV2) ListWatcher(ctx context.Context, req *api.ListWatcherRequest) (*api.ListWatcherResponse, error) {
-	xs, token, err := w.std.list(req)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &api.ListWatcherResponse{
-		NextPageToken: token,
-	}
-	for _, x := range xs {
-		v, err := w.revert(x)
-		if err != nil {
-			return nil, err
-		}
-		res.Watchers = append(res.Watchers, v)
-	}
-	return res, nil
-}
-
-func (w *WatcherV2) UpdateWatcher(ctx context.Context, req *api.UpdateWatcherRequest) (*api.Watcher, error) {
-	res, err := w.std.update(
-		func() error {
-			return nil
-		},
-		req,
-		req.GetWatcher(),
-	)
-	if err != nil {
-		return nil, err
-	}
-	return w.revert(res)
-}
-
-func (w *WatcherV2) DeleteWatcher(ctx context.Context, req *api.DeleteWatcherRequest) (*empty.Empty, error) {
-	return w.std.delete(
-		func() error {
-			return nil
-		},
-		req,
-	)
 }
